Move Dimension from compoundgenerator.py to new dimension.py
It is now expected for this class to become public API (after some
changes that are not part of this commit)
